==> 2_data_cleaning/party/1_get_each_party.ipynb <==
# 각 주별 파티데이터만 추출하기

import pandas as pd
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

## 훈련 데이터 추출

data = pd.read_csv("train_party.csv")

data.head()

len(data)

#각주별 데이터만을 뽑는다.
party_1wk = data[data.party_start_week == 1]
party_2wk = data[data.party_start_week == 2]
party_3wk = data[data.party_start_week == 3]
party_4wk = data[data.party_start_week == 4]
party_5wk = data[data.party_start_week == 5]
party_6wk = data[data.party_start_week == 6]
party_7wk = data[data.party_start_week == 7]
party_8wk = data[data.party_start_week == 8]

#주차별 데이터의 길이 확인
len(party_1wk)
len(party_2wk)
len(party_3wk)
len(party_4wk)
len(party_5wk)
len(party_6wk)
len(party_7wk)
len(party_8wk)

### 데이터 저장

party_1wk.to_csv("train_party_1wk.csv", index=False)
party_2wk.to_csv("train_party_2wk.csv", index=False)
party_3wk.to_csv("train_party_3wk.csv", index=False)
party_4wk.to_csv("train_party_4wk.csv", index=False)
party_5wk.to_csv("train_party_5wk.csv", index=False)
party_6wk.to_csv("train_party_6wk.csv", index=False)
party_7wk.to_csv("train_party_7wk.csv", index=False)
party_8wk.to_csv("train_party_8wk.csv", index=False)

## 평가 데이터 추출

data = pd.read_csv("test_party.csv")

party_1wk = data[data.party_start_week == 1]
party_2wk = data[data.party_start_week == 2]
party_3wk = data[data.party_start_week == 3]
party_4wk = data[data.party_start_week == 4]
party_5wk = data[data.party_start_week == 5]
party_6wk = data[data.party_start_week == 6]
party_7wk = data[data.party_start_week == 7]
party_8wk = data[data.party_start_week == 8]

party_1wk.to_csv("test_party_1wk.csv", index=False)
party_2wk.to_csv("test_party_2wk.csv", index=False)
party_3wk.to_csv("test_party_3wk.csv", index=False)
party_4wk.to_csv("test_party_4wk.csv", index=False)
party_5wk.to_csv("test_party_5wk.csv", index=False)
party_6wk.to_csv("test_party_6wk.csv", index=False)
party_7wk.to_csv("test_party_7wk.csv", index=False)
party_8wk.to_csv("test_party_8wk.csv", index=False)
