==> 2_data_cleaning/party/4_member_count.ipynb <==
# 각 파티에 대한 인원수 구하기
#### 각 주차당 동일 작업을 했습니다. 주석은 1주차에 해당하는 부분에만 달려있습니다.

import pandas as pd
import numpy as np

#1주차에 해당하는 파티데이터 불러오기
party_raw = pd.read_csv('train_party_1wk.csv')

#파티데이터 간단히 확인
party_raw.head(10)

#각 ID를 세서 인원수 확인
count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

#파티별 인원수 저장
save.to_csv("train_party_1wk_count.csv", index=False)

party_raw = pd.read_csv('train_party_2wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("train_party_2wk_count.csv", index=False)

party_raw = pd.read_csv('train_party_3wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("train_party_3wk_count.csv", index=False)

party_raw = pd.read_csv('train_party_4wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("train_party_4wk_count.csv", index=False)

party_raw = pd.read_csv('train_party_5wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("train_party_5wk_count.csv", index=False)

party_raw = pd.read_csv('train_party_6wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("train_party_6wk_count.csv", index=False)

party_raw = pd.read_csv('train_party_7wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("train_party_7wk_count.csv", index=False)

party_raw = pd.read_csv('train_party_8wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("train_party_8wk_count.csv", index=False)

party_raw = pd.read_csv('test_party_1wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("test_party_1wk_count.csv", index=False)

party_raw = pd.read_csv('test_party_2wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("test_party_2wk_count.csv", index=False)

party_raw = pd.read_csv('test_party_3wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("test_party_3wk_count.csv", index=False)

party_raw = pd.read_csv('test_party_4wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("test_party_4wk_count.csv", index=False)

party_raw = pd.read_csv('test_party_5wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("test_party_5wk_count.csv", index=False)

party_raw = pd.read_csv('test_party_6wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("test_party_6wk_count.csv", index=False)

party_raw = pd.read_csv('test_party_7wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("test_party_7wk_count.csv", index=False)

party_raw = pd.read_csv('test_party_8wk.csv')

party_raw.head(10)

len(party_raw)

count = []
for row in party_raw['hashed']:
    cnt = len(row.strip().split(','))
    count.append(cnt)
party_raw['count'] = count
save = pd.DataFrame(count, columns=['count'])

len(save)

save.to_csv("test_party_8wk_count.csv", index=False)


len(party_raw)
