==> 2_data_cleaning/party/5_union_time_member_10min.ipynb <==
# 각 파티에 대한 인원수, 시간 정보 합치기
## 파티 시간이 10분부터 의미있다고 간주.
#### 각 주차당 동일 작업을 했습니다. 주석은 1주차에 해당하는 부분에만 달려있습니다.

import pandas as pd
import os
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

os.chdir('../data')

#시간 정보 가진 데이터 불러오기
train_party_1wk_pre = pd.read_csv("train_party_1wk_time.csv")
#파티 인원수 데이터 불러오기
count = pd.read_csv("train_party_1wk_count.csv")

train_party_1wk_pre.head()
count.head()

len(train_party_1wk_pre)
len(count)

#데이터 합치기
train_party_1wk_pre["count"] = count
len(train_party_1wk_pre)

#파티 타임 10분 이상만 남긴다.
train_party_1wk_pre = train_party_1wk_pre[train_party_1wk_pre.time>10]

len(train_party_1wk_pre)

#데이터 저장
train_party_1wk_pre.to_csv("train_party_1wk_pre.csv", index = False)

train_party_1wk_pre = pd.read_csv("train_party_2wk_time.csv")
count = pd.read_csv("train_party_2wk_count.csv")

train_party_1wk_pre.head()
count.head()

len(train_party_1wk_pre)
len(count)

train_party_1wk_pre["count"] = count
len(train_party_1wk_pre)

train_party_1wk_pre = train_party_1wk_pre[train_party_1wk_pre.time>10]

len(train_party_1wk_pre)

train_party_1wk_pre.to_csv("train_party_2wk_pre.csv", index = False)

train_party_1wk_pre = pd.read_csv("train_party_3wk_time.csv")
count = pd.read_csv("train_party_3wk_count.csv")

train_party_1wk_pre.head()
count.head()

len(train_party_1wk_pre)
len(count)

train_party_1wk_pre["count"] = count
len(train_party_1wk_pre)

train_party_1wk_pre = train_party_1wk_pre[train_party_1wk_pre.time>10]

len(train_party_1wk_pre)

train_party_1wk_pre.to_csv("train_party_3wk_pre.csv", index = False)

train_party_1wk_pre = pd.read_csv("train_party_4wk_time.csv")
count = pd.read_csv("train_party_4wk_count.csv")

train_party_1wk_pre.head()
count.head()

len(train_party_1wk_pre)
len(count)

train_party_1wk_pre["count"] = count
len(train_party_1wk_pre)

train_party_1wk_pre = train_party_1wk_pre[train_party_1wk_pre.time>10]

len(train_party_1wk_pre)

train_party_1wk_pre.to_csv("train_party_4wk_pre.csv", index = False)

train_party_1wk_pre = pd.read_csv("train_party_5wk_time.csv")
count = pd.read_csv("train_party_5wk_count.csv")

train_party_1wk_pre.head()
count.head()

len(train_party_1wk_pre)
len(count)

train_party_1wk_pre["count"] = count
len(train_party_1wk_pre)

train_party_1wk_pre = train_party_1wk_pre[train_party_1wk_pre.time>10]

len(train_party_1wk_pre)

train_party_1wk_pre.to_csv("train_party_5wk_pre.csv", index = False)

train_party_1wk_pre = pd.read_csv("train_party_6wk_time.csv")
count = pd.read_csv("train_party_6wk_count.csv")

train_party_1wk_pre.head()
count.head()

len(train_party_1wk_pre)
len(count)

train_party_1wk_pre["count"] = count
len(train_party_1wk_pre)

train_party_1wk_pre = train_party_1wk_pre[train_party_1wk_pre.time>10]

len(train_party_1wk_pre)

train_party_1wk_pre.to_csv("train_party_6wk_pre.csv", index = False)

train_party_1wk_pre = pd.read_csv("train_party_7wk_time.csv")
count = pd.read_csv("train_party_7wk_count.csv")

train_party_1wk_pre.head()
count.head()

len(train_party_1wk_pre)
len(count)

train_party_1wk_pre["count"] = count
len(train_party_1wk_pre)

train_party_1wk_pre = train_party_1wk_pre[train_party_1wk_pre.time>10]

len(train_party_1wk_pre)

train_party_1wk_pre.to_csv("train_party_7wk_pre.csv", index = False)

train_party_1wk_pre = pd.read_csv("train_party_8wk_time.csv")
count = pd.read_csv("train_party_8wk_count.csv")

train_party_1wk_pre.head()
count.head()

len(train_party_1wk_pre)
len(count)

train_party_1wk_pre["count"] = count
len(train_party_1wk_pre)

train_party_1wk_pre = train_party_1wk_pre[train_party_1wk_pre.time>10]

len(train_party_1wk_pre)

train_party_1wk_pre.to_csv("train_party_8wk_pre.csv", index = False)

test_party_1wk_pre = pd.read_csv("test_party_1wk_time.csv")
count = pd.read_csv("test_party_1wk_count.csv")

test_party_1wk_pre.head()
count.head()

len(test_party_1wk_pre)
len(count)

test_party_1wk_pre["count"] = count
len(test_party_1wk_pre)

test_party_1wk_pre = test_party_1wk_pre[test_party_1wk_pre.time>10]

len(test_party_1wk_pre)

test_party_1wk_pre.to_csv("test_party_1wk_pre.csv", index = False)

test_party_1wk_pre = pd.read_csv("test_party_2wk_time.csv")
count = pd.read_csv("test_party_2wk_count.csv")

test_party_1wk_pre.head()
count.head()

len(test_party_1wk_pre)
len(count)

test_party_1wk_pre["count"] = count
len(test_party_1wk_pre)

test_party_1wk_pre = test_party_1wk_pre[test_party_1wk_pre.time>10]

len(test_party_1wk_pre)

test_party_1wk_pre.to_csv("test_party_2wk_pre.csv", index = False)

test_party_1wk_pre = pd.read_csv("test_party_3wk_time.csv")
count = pd.read_csv("test_party_3wk_count.csv")

test_party_1wk_pre.head()
count.head()

len(test_party_1wk_pre)
len(count)

test_party_1wk_pre["count"] = count
len(test_party_1wk_pre)

test_party_1wk_pre = test_party_1wk_pre[test_party_1wk_pre.time>10]

len(test_party_1wk_pre)

test_party_1wk_pre.to_csv("test_party_3wk_pre.csv", index = False)

test_party_1wk_pre = pd.read_csv("test_party_4wk_time.csv")
count = pd.read_csv("test_party_4wk_count.csv")

test_party_1wk_pre.head()
count.head()

len(test_party_1wk_pre)
len(count)

test_party_1wk_pre["count"] = count
len(test_party_1wk_pre)

test_party_1wk_pre = test_party_1wk_pre[test_party_1wk_pre.time>10]

len(test_party_1wk_pre)

test_party_1wk_pre.to_csv("test_party_4wk_pre.csv", index = False)

test_party_1wk_pre = pd.read_csv("test_party_5wk_time.csv")
count = pd.read_csv("test_party_5wk_count.csv")

test_party_1wk_pre.head()
count.head()

len(test_party_1wk_pre)
len(count)

test_party_1wk_pre["count"] = count
len(test_party_1wk_pre)

test_party_1wk_pre = test_party_1wk_pre[test_party_1wk_pre.time>10]

len(test_party_1wk_pre)

test_party_1wk_pre.to_csv("test_party_5wk_pre.csv", index = False)

test_party_1wk_pre = pd.read_csv("test_party_6wk_time.csv")
count = pd.read_csv("test_party_6wk_count.csv")

test_party_1wk_pre.head()
count.head()

len(test_party_1wk_pre)
len(count)

test_party_1wk_pre["count"] = count
len(test_party_1wk_pre)

test_party_1wk_pre = test_party_1wk_pre[test_party_1wk_pre.time>10]

len(test_party_1wk_pre)

test_party_1wk_pre.to_csv("test_party_6wk_pre.csv", index = False)

test_party_1wk_pre = pd.read_csv("test_party_7wk_time.csv")
count = pd.read_csv("test_party_7wk_count.csv")

test_party_1wk_pre.head()
count.head()

len(test_party_1wk_pre)
len(count)

test_party_1wk_pre["count"] = count
len(test_party_1wk_pre)

test_party_1wk_pre = test_party_1wk_pre[test_party_1wk_pre.time>10]

len(test_party_1wk_pre)

test_party_1wk_pre.to_csv("test_party_7wk_pre.csv", index = False)

test_party_1wk_pre = pd.read_csv("test_party_8wk_time.csv")
count = pd.read_csv("test_party_8wk_count.csv")

test_party_1wk_pre.head()
count.head()

len(test_party_1wk_pre)
len(count)

test_party_1wk_pre["count"] = count
len(test_party_1wk_pre)

test_party_1wk_pre = test_party_1wk_pre[test_party_1wk_pre.time>10]

len(test_party_1wk_pre)

test_party_1wk_pre.to_csv("test_party_8wk_pre.csv", index = False)
