==> 2_data_cleaning/party/6_sort_party.ipynb <==
# 뭉쳐져 있는 ID들 각 아이디 별로 풀어헤치기
#### 각 주차당 동일 작업을 했습니다. 주석은 1주차에 해당하는 부분에만 달려있습니다.

import pandas as pd
import numpy as np
import tqdm

#해시 뭉쳐져 있는 데이터 불러오기
df=pd.read_csv('train_party_1wk_pre.csv')

n=len(df)
#메모리관리와 속도를 위한 제네레이터 사용
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

#각 아이디별로 time, count 정보 담기
def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

#아이디, 시간, 멤버수에 따라 정렬
df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

#정렬된 데이터 저장
df.to_csv('train_party_1wk_sort.csv',index=False)

df=pd.read_csv('train_party_2wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('train_party_2wk_sort.csv',index=False)

df=pd.read_csv('train_party_3wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('train_party_3wk_sort.csv',index=False)

df=pd.read_csv('train_party_4wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('train_party_4wk_sort.csv',index=False)

df=pd.read_csv('train_party_5wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('train_party_5wk_sort.csv',index=False)

df=pd.read_csv('train_party_6wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('train_party_6wk_sort.csv',index=False)

df=pd.read_csv('train_party_7wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('train_party_7wk_sort.csv',index=False)

df=pd.read_csv('train_party_8wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('train_party_8wk_sort.csv',index=False)

#df=pd.read_csv('0828/party_test_0828.csv')
df=pd.read_csv('test_party_1wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('test_party_1wk_sort.csv',index=False)

df=pd.read_csv('test_party_2wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('test_party_2wk_sort.csv',index=False)

df=pd.read_csv('test_party_3wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('test_party_3wk_sort.csv',index=False)

df=pd.read_csv('test_party_4wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('test_party_4wk_sort.csv',index=False)

df=pd.read_csv('test_party_5wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('test_party_5wk_sort.csv',index=False)

df=pd.read_csv('test_party_6wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('test_party_6wk_sort.csv',index=False)

df=pd.read_csv('test_party_7wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('test_party_7wk_sort.csv',index=False)

df=pd.read_csv('test_party_8wk_pre.csv')

n=len(df)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

del(df)

def untie():
    ttime=[]
    hhashed=[]
    ccount=[]
    for i in tqdm.trange(n):
        tmphashed=next(hashed).split(',')
        hhashed+=tmphashed
        tmptime=next(ptime)
        tmpcount=next(count)
        for j in tmphashed:
            ttime.append(tmptime)
            ccount.append(tmpcount)
    res=pd.DataFrame({'time':ttime,'count':ccount,'hashed':hhashed})
    return res

df=untie()

df=df.sort_values(by=['hashed','time','count'])

len(df)

df[0:100]

df.to_csv('test_party_8wk_sort.csv',index=False)


alpha=list(chain.from_iterable(pd.read_csv('0828/party_train_0828.csv')['hashed'].str.split(',')))

alpha=list(set(alpha))
