==> 2_data_cleaning/party/7_get_party_data.ipynb <==
# 각 아이디에 해당 파티정보들 합치기, 가입했던 파티의 평균 멤버수 구하기
#### 각 주차당 동일 작업을 했습니다. 주석은 1주차에 해당하는 부분에만 달려있습니다.

import pandas as pd
import numpy as np
import tqdm

#정렬된 파티데이터 불러오기
df=pd.read_csv('train_party_1wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

#유일한 해시 정보
unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
#메모리관리, 속도를 위한 제네레이터 사용
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

#한 아이디가 여러개의 파티에 가입한 경우 그 정보를 합쳐준다.
def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

#가입했던 파티의 총 멤버수와 가입한 파티의 갯수를 나눈 평균 파티원 수
df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

#최종 파티 정보 데이터 저장
df.to_csv('train_party_1wk.csv',index=False)

df=pd.read_csv('train_party_2wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('train_party_2wk.csv',index=False)

df=pd.read_csv('train_party_3wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('train_party_3wk.csv',index=False)

df=pd.read_csv('train_party_4wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('train_party_4wk.csv',index=False)

df=pd.read_csv('train_party_5wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('train_party_5wk.csv',index=False)

df=pd.read_csv('train_party_6wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('train_party_6wk.csv',index=False)

df=pd.read_csv('train_party_7wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('train_party_7wk.csv',index=False)

df=pd.read_csv('train_party_8wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('train_party_8wk.csv',index=False)

df=pd.read_csv('test_party_1wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('test_party_1wk.csv',index=False)

df=pd.read_csv('test_party_2wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('test_party_2wk.csv',index=False)

df=pd.read_csv('test_party_3wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('test_party_3wk.csv',index=False)

df=pd.read_csv('test_party_4wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('test_party_4wk.csv',index=False)

df=pd.read_csv('test_party_5wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('test_party_5wk.csv',index=False)

df=pd.read_csv('test_party_6wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('test_party_6wk.csv',index=False)

df=pd.read_csv('test_party_7wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('test_party_7wk.csv',index=False)

df=pd.read_csv('test_party_8wk_sort.csv')
dfn=len(df)

dfn

#df[0:20]

unique_h=list(set(df['hashed']))
unique_h.sort()

#unique_h=np.load('party_id.npy')

unique_h

n=len(unique_h)
u_hashed=(row for row in unique_h)
hashed=(row for row in df['hashed'])
ptime=(row for row in df['time'])
count=(row for row in df['count'])

def hashsum():
    h=[None]*n
    t=[0]*n
    c=[0]*n
    pc=[0]*n
    tmp_uh=next(u_hashed)
    tmp_hashed=tmp_uh
    j=0
    for i in tqdm.trange(n):
        while(tmp_uh==tmp_hashed):
            if j==dfn:
                break
            tmp_hashed=next(hashed)
            h[i]=tmp_uh
            if tmp_hashed==tmp_uh:
                t[i]+=next(ptime)
                c[i]+=next(count)
                pc[i]+=1
            else:
                t[i+1]+=next(ptime)
                c[i+1]+=next(count)
                pc[i+1]+=1
            j+=1
        if i<n-1:
            tmp_uh=next(u_hashed)
    df=pd.DataFrame({'acc_id':h,'time':t,'member_count':c,'party_count':pc})
    return df

df=hashsum()

df

df['mean_mem']=df['member_count']/df['party_count']

len(df)
df.head()

df.to_csv('test_party_8wk.csv',index=False)
