==> 2_data_cleaning/party/8_merge_party_wk.ipynb <==
import pandas as pd
import os

os.chdir("../data")
tr_party1=pd.read_csv('party/train_party_1wk.csv')
tr_party2=pd.read_csv('party/train_party_2wk.csv')
tr_party3=pd.read_csv('party/train_party_3wk.csv')
tr_party4=pd.read_csv('party/train_party_4wk.csv')
tr_party5=pd.read_csv('party/train_party_5wk.csv')
tr_party6=pd.read_csv('party/train_party_6wk.csv')
tr_party7=pd.read_csv('party/train_party_7wk.csv')
tr_party8=pd.read_csv('party/train_party_8wk.csv')
te_party1=pd.read_csv('party/test_party_1wk.csv')
te_party2=pd.read_csv('party/test_party_2wk.csv')
te_party3=pd.read_csv('party/test_party_3wk.csv')
te_party4=pd.read_csv('party/test_party_4wk.csv')
te_party5=pd.read_csv('party/test_party_5wk.csv')
te_party6=pd.read_csv('party/test_party_6wk.csv')
te_party7=pd.read_csv('party/test_party_7wk.csv')
te_party8=pd.read_csv('party/test_party_8wk.csv')

# activity 데이터와 통합하기 위해 wk 설명변수를 만들어준다.
tr_party1['wk']=1
tr_party2['wk']=2
tr_party3['wk']=3
tr_party4['wk']=4
tr_party5['wk']=5
tr_party6['wk']=6
tr_party7['wk']=7
tr_party8['wk']=8
te_party1['wk']=1
te_party2['wk']=2
te_party3['wk']=3
te_party4['wk']=4
te_party5['wk']=5
te_party6['wk']=6
te_party7['wk']=7
te_party8['wk']=8

tr_party = pd.concat([tr_party1, tr_party2, tr_party3, tr_party4, tr_party5, tr_party6, tr_party7, tr_party8]).sort_values(by=['acc_id','wk'])
te_party = pd.concat([te_party1, te_party2, te_party3, te_party4, te_party5, te_party6, te_party7, te_party8]).sort_values(by=['acc_id','wk'])

tr_party.to_csv('train_party_pre.csv', index=False)
te_party.to_csv('test_party_pre.csv', index=False)
